backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "google-auth>=2.49.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "litellm[google]>=1.82.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
